EffectUIValidator, EffectClientInterface moved out of libraries
