+++ libraries/lib-components/EffectInterface.cpp
@@ -213,39 +213,6 @@ size_t EffectInstanceWithBlockSize::SetBlockSize(size_t maxBlockSize)
 
 EffectInstanceFactory::~EffectInstanceFactory() = default;
 
-EffectUIValidator::EffectUIValidator(
-   EffectUIClientInterface &effect, EffectSettingsAccess &access)
-   : mEffect{effect}
-   , mAccess{access}
-{}
-
-EffectUIValidator::~EffectUIValidator() = default;
-
-bool EffectUIValidator::UpdateUI()
-{
-   return true;
-}
-
-bool EffectUIValidator::IsGraphicalUI()
-{
-   return false;
-}
-
-void EffectUIValidator::Disconnect()
-{
-}
-
-void EffectUIValidator::OnClose()
-{
-   if (!mUIClosed)
-   {
-      mEffect.CloseUI();
-      mUIClosed = true;
-   }
-}
-
-EffectUIClientInterface::~EffectUIClientInterface() = default;
-
 const RegistryPath &CurrentSettingsGroup()
 {
    static RegistryPath id{ "CurrentSettings" };
